feat: #3- 뷰 파일 분리 MovieDetailView 에서 MovieDetailInfoView 분리
